requirements.txt
tensorflow
keras
numpy
matplotlib
opencv-python
pillow
streamlit
